Preliminary support for container prefixes
Docker compose calls these "projects" and uses them to namespace
containers. By default dev will change the prefix to the directory
where its configuration file is found or it can be specified in
the yaml with the container_prefix option. Hopefully can make this
project specific in the future.

This fixes #15.